+++ tests/multi_cluster/control/test_pod_configs.py
@@ -1,0 +1,8 @@
+"""
+Tests for the unified cluster scenario.
+In this scenario, all components (control plane and data plane) are installed in a single cluster.
+"""
+
+
+def test_dummy(control):
+    assert True
